Implemented matching users by identifying the biggest distance between to points in n-dimensional space. Room creation is limited to the farthest point and it needs to be revised to support multiple chats, and avoid repetition of matches. Chat support needs to be added now.

+++ Polink/View Controllers/QuizVC.swift
@@ -113,7 +113,7 @@ class QuizVC: UIViewController {
 	func startQuiz () {
 		quiz.initQuiz()
 		loadQuestion()
-		var delayCounter:Double = 1
+		var delayCounter:Float = 1
 		for button in buttons {
 			animateIn(button, delay: 1 + delayCounter)
 			delayCounter += 0.5
